refact: cli change dep package, fire change to click fix[app]: pydantic not allow model_config as class member variable fix[enode]: map /root/.config/vllm into container
